fix(inqaccu): truncate overdraft to integer pounds; close cursor on any throw
The DB column overdraft_limit is NUMERIC(12,2) but the COBOL/DTO contract
treats it as integer pounds; longValueExact() would turn any fractional
value into a 500. Use toBigInteger().longValueExact() to truncate the
fractional part while still failing fast on overflow.

Wrap the cursor fetch loop in try/finally so a non-DataAccessException
(e.g., a mapping/validation error) cannot bypass cursor close and leak
the underlying jOOQ cursor/connection.

+++ src/main/java/com/augment/cbsa/web/inqacccu/InqacccuController.java
@@ -105,7 +105,7 @@ public class InqacccuController {
                 account.accountType(),
                 account.interestRate(),
                 toCobolDate(account.opened()),
-                account.overdraftLimit().longValueExact(),
+                account.overdraftLimit().toBigInteger().longValueExact(),
                 toCobolDate(account.lastStatementDate()),
                 toCobolDate(account.nextStatementDate()),
                 account.availableBalance(),
